# docs/openAPI_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Expense SaaS API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tenants": {
      "post": {
        "tags": [
          "tenants"
        ],
        "summary": "Create Tenant",
        "description": "Create a new tenant and grant owner membership to the creator.\n\nArgs:\n    payload: TenantCreate input schema with the tenant name.\n    db: Async DB session.\n    user: Current authenticated user (will become tenant owner).\n\nReturns:\n    The created Tenant object.",
        "operationId": "create_tenant_tenants_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "tenants"
        ],
        "summary": "List Tenants",
        "description": "List tenants the current user is an active member of.\n\nArgs:\n    db: Async DB session.\n    user: Current authenticated user.\n\nReturns:\n    List of Tenant objects the user is a member of.",
        "operationId": "list_tenants_tenants_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TenantRead"
                  },
                  "title": "Response List Tenants Tenants Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tenants/{tenant_id}": {
      "get": {
        "tags": [
          "tenants"
        ],
        "summary": "Get Tenant",
        "description": "Retrieve a tenant by id if the requester is an active member.\n\nArgs:\n    tenant_id: UUID of the tenant to retrieve.\n    db: Async DB session.\n    user: Current authenticated user.\n\nReturns:\n    The Tenant object when the user is an active member.\n\nRaises:\n    HTTPException 403 when the user is not a member.\n    HTTPException 404 when the tenant does not exist.",
        "operationId": "get_tenant_tenants__tenant_id__get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "tenants"
        ],
        "summary": "Update Tenant",
        "description": "Update tenant properties; only owners may perform this action.\n\nArgs:\n    tenant_id: UUID of the tenant to update.\n    payload: TenantUpdate schema with changes.\n    db: Async DB session.\n    user: Current authenticated user; must be OWNER.\n\nReturns:\n    The updated Tenant object.\n\nRaises:\n    HTTPException 403 when the user is not the owner.\n    HTTPException 404 when the tenant does not exist.",
        "operationId": "update_tenant_tenants__tenant_id__patch",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tenants"
        ],
        "summary": "Delete Tenant",
        "description": "Delete a tenant. Only owners can delete their tenant.\n\nArgs:\n    tenant_id: UUID of the tenant to delete.\n    db: Async DB session.\n    user: Current authenticated user; must be OWNER.\n\nRaises:\n    HTTPException 403 when the user is not the owner.\n    HTTPException 404 when the tenant does not exist.",
        "operationId": "delete_tenant_tenants__tenant_id__delete",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tenants/{tenant_id}/switch": {
      "post": {
        "tags": [
          "tenants"
        ],
        "summary": "Switch Active Tenant",
        "description": "Issue a new access token with the requested tenant_id after validating membership.\n\nReturns:\n    {\"access_token\": \"\u003Cjwt\u003E\"}\nRaises:\n    403 if user is not an active member of the requested tenant.\n    404 if tenant not found.",
        "operationId": "switch_active_tenant_tenants__tenant_id__switch_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Switch Active Tenant Tenants  Tenant Id  Switch Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tenants/{tenant_id}/members": {
      "post": {
        "tags": [
          "tenants"
        ],
        "summary": "Create Membership For Tenant",
        "description": "Create a membership or invite for a tenant. Only tenant owners may invite.",
        "operationId": "create_membership_for_tenant_tenants__tenant_id__members_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MembershipCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "tenants"
        ],
        "summary": "List Members For Tenant",
        "description": "List all memberships for a tenant. Requester must be an active member.",
        "operationId": "list_members_for_tenant_tenants__tenant_id__members_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MembershipRead"
                  },
                  "title": "Response List Members For Tenant Tenants  Tenant Id  Members Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MembershipCreate": {
        "properties": {
          "user_email": {
            "type": "string",
            "title": "User Email"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MembershipRole"
              },
              {
                "type": "null"
              }
            ],
            "default": "member"
          }
        },
        "type": "object",
        "required": [
          "user_email"
        ],
        "title": "MembershipCreate",
        "description": "Input schema to create a membership or invite within a tenant.\nTenant ID is derived from the active context within the access_token\n\nArgs:\n    user_email: Email of the user to invite or add.\n    role: Role to assign to the membership."
      },
      "MembershipRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "user_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Email"
          },
          "role": {
            "$ref": "#/components/schemas/MembershipRole"
          },
          "status": {
            "$ref": "#/components/schemas/MembershipStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "user_id",
          "user_email",
          "role",
          "status",
          "created_at"
        ],
        "title": "MembershipRead",
        "description": "Read schema returning membership details.\nTenant ID is derived from the active context within the access_token\n\nAttributes:\n    id: Membership identifier.\n    tenant_id: Tenant the membership belongs to.\n    user_id: Optional linked user id.\n    user_email: Optional invite email.\n    role: Assigned role.\n    status: Membership status.\n    created_at: Creation timestamp."
      },
...
      "TenantCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "TenantCreate",
        "description": "Input schema to create a new tenant.\n\nArgs:\n    name: Human-readable name for the tenant."
      },
      "TenantRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "title": "TenantRead",
        "description": "Read schema for tenant information returned by the API.\n\nAttributes:\n    id: Tenant identifier.\n    name: Tenant name.\n    created_at: Creation timestamp."
      },
      "TenantUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "title": "TenantUpdate",
        "description": "Input schema for updating a tenant's properties.\n\nArgs:\n    name: New name for the tenant (optional)."
      },
...
    }
  }
}